+++ src/main/java/com/kanghwang/khholdings/domain/market/MarketRepository.java
@@ -6,17 +6,17 @@ import java.util.List;
 import com.kanghwang.khholdings.domain.order.dto.CandleDTO;
 import org.apache.ibatis.annotations.Mapper;
 
-import com.kanghwang.khholdings.domain.market.dto.MarketDTO;
+import com.kanghwang.khholdings.domain.market.dto.TokenListDTO;
 import com.kanghwang.khholdings.domain.market.dto.PendingDTO;
 
 @Mapper
 public interface MarketRepository {
 
 	// 종목 전체 조회
-	List<MarketDTO> selectAll();
+	List<TokenListDTO> selectAll();
 
 	// 종목 검색어 조회
-	List<MarketDTO> selectBySearch(String content);
+	List<TokenListDTO> selectBySearch(String content);
 
 	// 미체결 내역 조회
 	List<PendingDTO> selectPending(Long tokenId, Long walletId);
