+++ src/main/java/com/kanghwang/khholdings/domain/order/TradeWorker.java
@@ -3,7 +3,6 @@ package com.kanghwang.khholdings.domain.order;
 import java.math.BigDecimal;
 import java.time.Duration;
 import java.time.OffsetDateTime;
-import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -12,9 +11,6 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
-import org.redisson.api.BatchResult;
-import org.redisson.api.RBatch;
-import org.redisson.api.RScript;
 import org.redisson.api.RStream;
 import org.redisson.api.RedissonClient;
 import org.redisson.api.StreamMessageId;
@@ -23,8 +19,8 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
+import com.kanghwang.khholdings.domain.market.Service.MarketDataService;
 import com.kanghwang.khholdings.domain.my.dto.TransactionHistDTO;
-import com.kanghwang.khholdings.domain.order.dto.CandleDTO;
 import com.kanghwang.khholdings.domain.order.dto.RefundRequestDTO;
 import com.kanghwang.khholdings.domain.order.dto.SettlementResultDTO;
 import com.kanghwang.khholdings.domain.order.dto.TransactionRequestDTO;
@@ -44,28 +40,10 @@ public class TradeWorker implements CommandLineRunner {
     private volatile boolean isRunning = true;
     private final CountDownLatch shutdownLatch = new CountDownLatch(1); // 종료 확인을 위한 래치 (1개의 스레드가 끝날 때까지 대기)
     private final RedisKeyManager redisKeyManager;
+	private final MarketDataService  marketDataService;
 
     // 체결 내역 벌크 인서트를 위한 버퍼
     private final Queue<TransactionHistDTO> transactionBuffer = new ConcurrentLinkedQueue<>();
-
-    // 1분 봉 업데이트 용 Lua Script
-    private static final String luaScript = """
-            local c = redis.call('HMGET', KEYS[1], 'high', 'low', 'close', 'vol')
-            local price = tonumber(ARGV[1])
-            local vol = tonumber(ARGV[2])
-
-            if not c[1] then
-                -- 새 봉 생성
-                redis.call('HMSET', KEYS[1], 'open', price, 'high', price, 'low', price, 'close', price, 'vol', vol)
-            else
-                -- 기존 봉 업데이트
-                if price > tonumber(c[1]) then redis.call('HSET', KEYS[1], 'high', price) end
-                if price < tonumber(c[2]) then redis.call('HSET', KEYS[1], 'low', price) end
-                redis.call('HSET', KEYS[1], 'close', price)
-                redis.call('HINCRBYFLOAT', KEYS[1], 'vol', vol)
-            end
-            redis.call('EXPIRE', KEYS[1], 10800) -- 3시간 TTL
-            """;
 
     @Override
     public void run(String... args) throws Exception {
@@ -153,8 +131,8 @@ public class TradeWorker implements CommandLineRunner {
             SettlementResultDTO result = new SettlementResultDTO();
             orderRepository.p_process_transaction_settlement(trade, result);
 
-            // 2. [DB] DB 기록 용 1분 봉 제작
-            updateRedisCandle(trade);
+            // 2. 1분 봉 제작, 토큰 실시간 리스트 제작
+			marketDataService.processMarketUpdate(trade);
 
             // 3. [비동기 기록] 거래 내역 로그 생성 및 버퍼 추가
             enqueueTransactionLogs(trade, result);
@@ -168,113 +146,7 @@ public class TradeWorker implements CommandLineRunner {
         }
     }
 
-    // [DB] 1분 봉 집계 로직
-    private void updateRedisCandle(TransactionRequestDTO trade) {
 
-        // 1분 단위로 버킷팅 (ex: 12:05:33 -> 12:05:00
-        long minute = trade.getCreatedAt().truncatedTo(ChronoUnit.MINUTES).toEpochSecond();
-        String candleKey = "candle:1m:" + trade.getTokenId() + ":" + minute;
-
-        redissonClient.getScript(org.redisson.client.codec.StringCodec.INSTANCE).eval(
-                RScript.Mode.READ_WRITE,
-                luaScript, // 상단에서 작성했던 거
-                RScript.ReturnType.VALUE,
-                List.of(candleKey),
-                trade.getTargetPrice().toPlainString(),
-                trade.getExecutedVolume().toPlainString());
-
-        // 업데이트된 최신 캔들 정보를 읽어서 전파
-        // 추후에 트레이딩뷰 차트를 실시간으로 움직이게 함
-        Map<String, String> candleMap = redissonClient
-                .<String, String>getMap(candleKey, org.redisson.client.codec.StringCodec.INSTANCE)
-                .readAllMap();
-
-        if (candleMap != null && !candleMap.isEmpty()) {
-            String topicKey = redisKeyManager.getCandleTopicKey(trade.getTokenId());
-
-            // 시간은 프론트엔드에서 한국 시간으로 변경 예정
-            CandleDTO liveCandle = CandleDTO.builder()
-                    .tokenId(trade.getTokenId())
-                    .unit(1)
-                    .candleTime(minute)
-                    .openingPrice(new BigDecimal(candleMap.get("open")))
-                    .highPrice(new BigDecimal(candleMap.get("high")))
-                    .lowPrice(new BigDecimal(candleMap.get("low")))
-                    .closingPrice(new BigDecimal(candleMap.get("close")))
-                    .tradeVolume(new BigDecimal(candleMap.get("vol")))
-                    .build();
-
-            // [MarketWoker - 차트]
-            // DTO 자체를 Redis Topic으로 발행 (MarketWorker가 받음)
-            redissonClient.getTopic(topicKey).publish(liveCandle);
-        }
-    }
-
-    // 매 분 5초에 실행
-    // 모으는 건 1분, 저장은 5초 뒤
-    @Scheduled(cron = "5 * * * * *")
-    public void syncCandleToDB() {
-        // 1. 방금 마감된 1분 계산
-        long lastMinute = OffsetDateTime.now().minusMinutes(1).truncatedTo(ChronoUnit.MINUTES).toEpochSecond();
-        String pattern = "candle:1m:*:" + lastMinute;
-
-        try {
-            // 2. 해당 시간내의 모든 토큰 키 찾기
-            Iterable<String> keys = redissonClient.getKeys().getKeysByPattern(pattern);
-
-            // 3. Redisson Batch 기능을 써서 한꺼번에 읽기 준비
-            RBatch batch = redissonClient.createBatch();
-            List<String> keyList = new ArrayList<>();
-
-            for (String key : keys) {
-                keyList.add(key);
-                // "나중에 이 키들 데이터 한꺼번에 읽어올 거야"라고 예약만 함
-                batch.getMap(key, org.redisson.client.codec.StringCodec.INSTANCE).readAllMapAsync();
-            }
-
-            if (keyList.isEmpty()) {
-                return;
-            }
-
-            // 4. 한 번의 네트워크 통신으로 모든 데이터 수집
-            BatchResult<?> result = batch.execute();
-            List<Map<String, String>> allRawData = (List<Map<String, String>>) result.getResponses();
-
-            // 5. 데이터 가공 (String -> BigDecimal & DTO화)
-            List<CandleDTO> candleList = new ArrayList<>();
-            for (int i = 0; i < keyList.size(); i++) {
-
-                Map<String, String> raw = allRawData.get(i);
-                if (raw == null || raw.isEmpty()) continue;
-
-                // 토큰 아이디 추출
-                String[] parts = keyList.get(i).split(":");
-                Long tokenId = Long.valueOf(parts[2]);
-
-                CandleDTO candle = CandleDTO.builder()
-                    .tokenId(tokenId)
-                    .unit(1)
-                    .candleTime(lastMinute)
-                    .openingPrice(new BigDecimal(raw.get("open")))
-                    .highPrice(new BigDecimal(raw.get("high")))
-                    .lowPrice(new BigDecimal(raw.get("low")))
-                    .closingPrice(new BigDecimal(raw.get("close")))
-                    .tradeVolume(new BigDecimal(raw.get("vol")))
-                    .build();
-
-                candleList.add(candle);
-            }
-
-            // 5. DB에 한꺼번에 저장
-            if (!candleList.isEmpty()) {
-                orderRepository.insertCandlesBatch(candleList);
-                log.info("[TradeWorker - Sync] {} 시점의 1분 봉 {}건을 DB로 저장 완료", lastMinute, candleList.size());
-            }
-
-        } catch (Exception e) {
-            log.error("[TradeWorker - Sync] 1분 봉 DB 동기화 중 오류 발생: ", e);
-        }
-    }
 
     // [DB] 체결 내역 생성
     // TransactionRequestDTO 누가 누구랑 얼마에 체결됐는가?
